requirements.txt
pillow
PyDrive